requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pinecone-client==3.0.0
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
langchain-community==0.0.13
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tiktoken==0.5.2
sqlalchemy==2.0.25
